[Fix #94] Fix circular dependencies
